pyproject.toml
[tool.poetry]
name = "doorstop-edit"
version = "0.1.0"
description = "Cross-platform doorstop GUI editor"
authors = ["Alexander Ernfridsson <ernfridsson@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "doorstop_edit"}]
include = [
     # override .gitignore
    "doorstop_edit/ui_gen/*_rc.py",
    "doorstop_edit/ui_gen/ui_*.py",
    "doorstop_edit/build_info.py"
]
exclude = [
    "*/**/test"
]
homepage = "https://github.com/ownbee/doorstop-edit"
repository = "https://github.com/ownbee/doorstop-edit"
keywords = ["doorstop", "requirements-management", "documentation", "gui"]
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Editors :: Documentation",
    "Topic :: Text Processing :: Markup",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
PySide6 = ">=6.4.0,<7.0.0"
pyqtdarktheme = "^2.1.0"
doorstop = "^2.2.post1"
mdformat = "^0.7.16"
mdformat_tables = "^0.4.1"
markdown-it-py = "^2.1.0"
qt-material = "^2.12"
pygments = "^2.14.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
black = "^22.12.0"
mypy = "^0.991"
types-requests = "^2.28.11.7"
flake8 = "^6.0.0"
isort = "^5.11.4"
colorama = "^0.4.6"
autoflake = "^2.0.0"


[tool.black]
line-length = 120
target-version = ['py310']

[tool.autoflake]
imports = ["requests", "PySide6"]

[tool.isort]
profile = "black"

[tool.poetry.scripts]
doorstop-edit = 'doorstop_edit.main:main'

[tool.poetry.build]
generate-setup-file = false
script = "build.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



